initialisation de la struct camera, remplacement de la structure color par int(get_rgba), et test init_sphere

+++ includes/camera.h
@@ -6,7 +6,7 @@
 /*   By: gpasquet <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/14 09:25:22 by gpasquet          #+#    #+#             */
-/*   Updated: 2023/04/14 10:18:20 by gpasquet         ###   ########.fr       */
+/*   Updated: 2023/04/17 16:32:06 by gpasquet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ typedef struct s_camera	t_camera;
 struct s_camera	{
 	t_coords	*pos;
 	t_vector	*orientation_vector;
-	int			fov;
+	double			fov;
 };
 
 #endif
